Time: 0 ms (100%), Space: 8.1 MB (84.87%) - LeetHub

+++ 0796-rotate-string/0796-rotate-string.cpp
@@ -17,7 +17,7 @@ public:
         }
 
         string doubleString = s+s;
-        if(doubleString.find(goal) < (int)doubleString.length()){
+        if(doubleString.find(goal) < doubleString.length()){
             return true;
         }
